# source/api/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is Process Wiz App.",
    "version": "1.0.0",
    "title": "Process Wiz"
  },
  "host": "api.pwdev.internal.nlightnconsulting.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/file-processor/custom-instruction": {
      "get": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to get custom instruction.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "collection_id",
            "description": "Collection id of the collection to get the custom instruction.",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      },
      "put": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to update custom instruction.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to update custom instruction.",
            "required": true,
            "schema": {
              "properties": {
                "collection_id": {
                  "type": "number"
                },
                "custom_instruction": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/file-processor/metadata/status": {
      "get": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to get status of metadata.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "collection_id",
            "description": "Collection id of the collection to get the metadata status.",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "total_files_count": {
                      "type": "number"
                    },
                    "total_processed_files_count": {
                      "type": "number"
                    },
                    "percentage_value": {
                      "type": "number"
                    },
                    "percentage_display_value": {
                      "type": "string"
                    }
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/file-processor/metadata": {
      "get": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to get content of metadata.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "collection_id",
            "description": "Collection id of the collection to get the metadata content.",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "file_id": {
                        "type": "number"
                      },
                      "file_name": {
                        "type": "string"
                      },
                      "processed_percentage": {
                        "type": "number"
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/file-processor/metadata/upload": {
      "post": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to upload csv file to update metadata. (File must be .csv)",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "description": "File to upload",
            "required": true,
            "type": "file"
          },
          {
            "name": "collection_id",
            "in": "formData",
            "description": "Id of the collection in which the file to upload",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/NormalResponse"
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/file-processor/summarise": {
      "get": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to get content of summarise.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "collection_id",
            "description": "Collection id of the collection to get the summarise content.",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "file_name": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "summary_short": {
                      "type": "string"
                    },
                    "summary_long": {
                      "type": "string"
                    }
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/file-processor/summarise/status": {
      "get": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to get status of summarise.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "collection_id",
            "description": "Collection id of the collection to get the summarise status.",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "total_files_count": {
                      "type": "number"
                    },
                    "total_processed_files_count": {
                      "type": "number"
                    },
                    "percentage_value": {
                      "type": "number"
                    },
                    "percentage_display_value": {
                      "type": "string"
                    }
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/file-processor/doc-guide/process-file": {
      "post": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to process doc guide files.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to process doc guide files",
            "required": true,
            "schema": {
              "properties": {
                "file_ids": {
                  "type": "array",
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      },
      "delete": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to delete doc guide files.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to process doc guide files",
            "required": true,
            "schema": {
              "properties": {
                "file_ids": {
                  "type": "array",
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/file-processor/metadata/update": {
      "post": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to update metadata.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to update metadata",
            "required": true,
            "schema": {
              "properties": {
                "collection_id": {
                  "type": "number"
                },
                "file_ids": {
                  "type": "array",
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
    "/file-processor/summarise/update": {
      "post": {
        "tags": [
          "FileProcessor"
        ],
        "description": "API to update metadata.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request body to update metadata",
            "required": true,
            "schema": {
              "properties": {
                "collection_id": {
                  "type": "number"
                },
                "is_process_remaining": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "summary_short": {
                        "type": "string"
                      },
                      "summary_long": {
                        "type": "string"
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "failure operation",
            "schema": {
              "$ref": "#/definitions/NormalFailureResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "organization-id": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Type in the *'Value'* input box below: **'Bearer &lt;JWT&gt;'**, where JWT is the token"
    },
    "organization-id": {
      "type": "apiKey",
      "in": "header",
      "name": "organization-id",
      "description": "Id of the organization"
    }
  },
  "definitions": {
...
    "NormalResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "NormalFailureResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
  }
}